app/requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docopt==0.6.2
Flask==1.1.1
html5lib==1.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
pipreqs==0.4.9
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
webencodings==0.5.1
Werkzeug==0.16.0
yarg==0.1.9
